pyproject.toml
[tool.poetry]
name = "notify"
version = "0.1.0"
description = ""
authors = ["Evgeny Kemerov <eskemerov@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.14.0"
psycopg2-binary = "^2.8"
sentry-sdk = {version = "^0.16.0", extras = ["celery"]}
django = "^3.0"
djangorestframework = "^3.11.0"
celery = {version = "^4.4.6", extras = ["redis"]}
uwsgi = "^2.0.19"
python-telegram-bot = "^12.8"

[tool.poetry.dev-dependencies]
pylint = "^2.3"
flake8 = "^3.7"
isort = "^5.1"
pylint-django = "^2.1"
black = "^19.10b0"
ipython = "^7.9.0"
mypy = "^0.782"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
